Refactor for efficient final set up
