+++ src/app/features/data-processing/data-processing.component.html
@@ -66,7 +66,7 @@
   <button
     (click)="onQuit()"
     mat-flat-button
-    style="background-color: #9333ea; color: white">
+    style="background-color: #9333ea; color: white; margin-top: 1rem">
     Back
   </button>
 </section>
