+++ src/app/layout/dashboard/dashboard.component.scss
@@ -59,7 +59,8 @@
 
         .logo {
           h1 {
-            font-weight: bold;
+            //font-weight: bold;
+            color: #f7f7f7;
             margin: 0;
           }
         }
@@ -111,87 +112,6 @@
 
       @media screen and (max-width: 768px) {
         margin: auto;
-      }
-
-      .dashboard-content {
-        margin: 1rem auto;
-        flex: 1;
-        display: flex;
-        justify-content: space-between;
-        gap: 2rem;
-
-        .center-card {
-          overflow-x: hidden;
-          padding: 0.3rem;
-          flex: 1;
-          background-color: #2a3448;
-          border-radius: 12px;
-          box-shadow: 0 0 20px rgba(0, 0, 0, 0.3);
-          border-color: #9333ea;
-        }
-
-        .side-card {
-          position: sticky;
-          top: 70px;
-          z-index: 1000;
-          border-color: #9333ea;
-          width: 20%;
-          background-color: #2a3448;
-          border-radius: 12px;
-          box-shadow: 0 0 20px rgba(0, 0, 0, 0.3);
-          max-height: 420px;
-          //max-height: 540px;
-
-          mat-divider {
-            height: 1px;
-            margin: 0.34rem;
-            background-color: #9333ea;
-          }
-
-          @media screen and (max-width: 1426px) {
-            min-height: 450px;
-          }
-
-          @media screen and (max-width: 1120px) {
-            min-height: 500px;
-          }
-
-          @media screen and (max-width: 1099px) {
-            min-height: 540px;
-          }
-
-          .mat-mdc-card-title {
-            padding: 0.8rem;
-
-            h2 {
-              font-weight: 600;
-              color: white;
-            }
-          }
-
-          .mat-mdc-card-content {
-            p {
-              color: white;
-            }
-          }
-
-          .mat-mdc-card-actions {
-            display: flex;
-            justify-content: flex-end;
-            //gap: 0.5rem;
-
-            button {
-              background-color: #9333ea;
-              color: white;
-            }
-          }
-        }
-
-        @media screen and (max-width: 1024px) {
-          .side-card {
-            display: none;
-          }
-        }
       }
     }
