+++ src/app/layout/landing-page/landing-page.component.html
@@ -1,0 +1,32 @@
+<!--        Summary Cards-->
+<ng-container>
+  <app-summary-card [summaries]="summaryData$" />
+</ng-container>
+
+<div class="dashboard-content">
+  <!--          Data Generation-->
+  <mat-card appearance="outlined" class="side-card left-card">
+    <mat-card-title><h2>Data Generation</h2></mat-card-title>
+    <mat-card-content>
+      <p>Generate dummy student data and export it to an excel file</p>
+      <p>Wanna give it a try?</p>
+    </mat-card-content>
+    <mat-card-actions>
+      <button (click)="navToDataGeneration()" mat-flat-button>Yes</button>
+    </mat-card-actions>
+    <mat-divider></mat-divider>
+    <mat-card-title><h2>Data Processing</h2></mat-card-title>
+    <mat-card-content>
+      <p>Upload Excel with student data and update scores with 10</p>
+      <p>Wanna give it a try?</p>
+    </mat-card-content>
+    <mat-card-actions>
+      <button (click)="navToDataProcessing()" mat-flat-button>Yes</button>
+    </mat-card-actions>
+  </mat-card>
+
+  <!--        Students Management-->
+  <mat-card appearance="outlined" class="center-card">
+    <app-students />
+  </mat-card>
+</div>
