+++ src/styles.scss
@@ -104,6 +104,7 @@ body {
   h2 {
     font-size: 1.5rem;
     font-weight: 400;
+    color: white;
   }
 
   .mat-mdc-card {
